requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
langchain-openai==0.0.5
tenacity==8.2.3 
gunicorn==22.10.0